test(auth): move tests to __tests__
