+++ backend/src/auth/__tests__/auth.controller.spec.ts
@@ -2,13 +2,12 @@ import { getModelToken } from '@nestjs/sequelize'
 import { Test, TestingModule } from '@nestjs/testing'
 import { PinoLogger } from 'nestjs-pino'
 
-import { ConfigModule } from '../config/config.module'
-import { User } from '../database/models'
-import { MailerModule } from '../mailer/mailer.module'
-import { OtpModule } from '../otp/otp.module'
-
-import { AuthController } from './auth.controller'
-import { AuthService } from './auth.service'
+import { ConfigModule } from '../../config/config.module'
+import { User } from '../../database/models'
+import { MailerModule } from '../../mailer/mailer.module'
+import { OtpModule } from '../../otp/otp.module'
+import { AuthController } from '../auth.controller'
+import { AuthService } from '../auth.service'
 
 describe('AuthController', () => {
   let controller: AuthController
